+++ objects/src/transaction/mod.rs
@@ -5,6 +5,7 @@ use super::{
     BlockHeader, Digest, Felt, Hasher, Word, WORD_SIZE, ZERO,
 };
 
+mod account_update;
 mod chain_mmr;
 mod executed_tx;
 mod inputs;
@@ -15,15 +16,13 @@ mod transaction_id;
 mod tx_args;
 mod tx_witness;
 
+pub use account_update::{AccountUpdate, AccountUpdateData, AccountUpdateDetails};
 pub use chain_mmr::ChainMmr;
 pub use executed_tx::ExecutedTransaction;
 pub use inputs::{InputNote, InputNotes, ToNullifier, TransactionInputs};
 pub use outputs::{OutputNote, OutputNotes, TransactionOutputs};
 pub use prepared_tx::PreparedTransaction;
-pub use proven_tx::{
-    AccountUpdate, AccountUpdateData, AccountUpdateDetails, ProvenTransaction,
-    ProvenTransactionBuilder,
-};
+pub use proven_tx::{ProvenTransaction, ProvenTransactionBuilder};
 pub use transaction_id::TransactionId;
 pub use tx_args::{TransactionArgs, TransactionScript};
 pub use tx_witness::TransactionWitness;