+++ CHANGELOG.md
@@ -3,8 +3,7 @@
 ## 0.3.0 (TBD)
 
 * [BREAKING] Removed the transaction script root output from the transaction kernel (#608).
-* [BREAKING] Refactored account update details (#618).
-* [BREAKING] Moved `Block` and `AccountUpdataData` to `miden-objects` (#621).
+* [BREAKING] Refactored account update details, moved `Block` to `miden-objects` (#618, #621).
 
 ## 0.2.1 (2024-04-12)